+++ src/deep/set.js
@@ -3,13 +3,6 @@
 import isNotEmptyString from '../string/isNotEmpty.js';
 import {PROTO_OBJ}      from '../object/is.js';
 
-//  Cleanup paths : a.b[2].c --> ['a', 'b', '2', 'c'] (faster processing)
-function interpolatePath (path) {
-    if (Array.isArray(path)) return [...path];
-    if (isNotEmptyString(path)) return path.replace('[', '.').replace(']', '').split('.');
-    throw new TypeError('No Path was given');
-}
-
 //  Set a value for a path in a json-like structure
 export default function deepSet (obj, path, value = null, define = false) {
     if (
@@ -17,7 +10,11 @@ export default function deepSet (obj, path, value = null, define = false) {
         !Array.isArray(obj)
     ) throw new TypeError('Deepset is only supported for objects');
 
-    const parts = interpolatePath(path);
+    //  If no path is provided, do nothing
+    if (!isNotEmptyString(path)) throw new TypeError('No path was given');
+
+    //  Cleanup paths : a.b[2].c --> ['a', 'b', '2', 'c'] (faster processing)
+    const parts = path.replace('[', '.').replace(']', '').split('.');
 
     //  Build any unknown paths and set cursor
     for (let i = 0; i < parts.length - 1; i++) {
@@ -28,16 +25,10 @@ export default function deepSet (obj, path, value = null, define = false) {
 
         //  Set cursor
         obj = obj[parts[i]];
-
-        //  If not an array continue
-        if (!Array.isArray(obj)) continue;
-
-        //  If an array and i is not the last index ... set the object to be the index on the array
-        if (i < parts.length - 2) {
-            obj = obj[parts[i + 1]];
-            i++;
-        }
     }
+
+    //  Prevent overriding of properties, eg: {d: 'hello'} -> deepSet('d.a.b', 'should not work')
+    if (!Array.isArray(obj) && Object.prototype.toString.call(obj) !== PROTO_OBJ) return false;
 
     //  Set the actual value on the cursor
     if (define) {